# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the professional backend API for the Echo messaging platform.",
        "title": "Echo Messenger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Moeed-ul-Hassan",
            "url": "http://www.moeed.me",
            "email": "moeed@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8001",
    "basePath": "/",
    "paths": {
...
        "/api/rooms/": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Create private room",
                "parameters": [
                    {
                        "description": "name, passkey, aes_key",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/room.Room"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/stories/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "stories"
                ],
                "summary": "List active stories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/story.Story"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stories"
                ],
                "summary": "Create story",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/story.Story"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "room.Room": {
            "type": "object",
            "properties": {
                "aes_key": {
                    "description": "E2EE Master Key (encrypted or base64)",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invite_only": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "require_2fa": {
                    "type": "boolean"
                },
                "room_id": {
                    "description": "Human-readable ID (e.g. \"Echo-1234\")",
                    "type": "string"
                },
                "ttl_hours": {
                    "type": "integer"
                }
            }
        },
        "story.Story": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "description": "Set to CreatedAt + 24h",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "media_type": {
                    "description": "image, video",
                    "type": "string"
                },
                "media_url": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "viewers": {
                    "description": "Users who swiped up/viewed",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/story.Viewer"
                    }
                }
            }
        },
        "story.Viewer": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string"
                },
                "viewed_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}